refact: move tabsName to data folder
